// src/lib/api/model/conditionCreate.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Unplan API
 * Unplan Server API Documentation
 * OpenAPI spec version: v1.0.0
 */

export interface ConditionCreate {
  /**
   * @minimum 0
   * @maximum 6
   */
  body_score: number;
  /**
   * @minimum 0
   * @maximum 6
   */
  mind_score: number;
  /** 기록 시각 */
  date_time: string;
}