// src/lib/api/model/conditionUpdate.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Unplan API
 * Unplan Server API Documentation
 * OpenAPI spec version: v1.0.0
 */

export interface ConditionUpdate {
  /**
   * @minimum 0
   * @maximum 6
   */
  body_score: number;
  /**
   * @minimum 0
   * @maximum 6
   */
  mind_score: number;
  /** 수정할 기록 시각 */
  date_time: string;
}